fix duration on failure
After commit d452782 (write end stage to stages file including the
accumulated build duration) the accumulated build duration is assumed to
be present in the stages file. If the build fails before reaching the
end stage, this assumption no longer holds. Therefore, always calculate
the build duration for correctness.

While here, do not try to calculate the delta if the previous release
failed.