tweak cvs log date
Using the time of the previous cvs stage when figuring out the new cvs
revisions turned out to be a bad idea. Some revisions might occurred
before this time if the cvs mirror is not fully up-to-date. Instead, use
the date from the last cvs revision in the previous release.