rework handling of release directories
Split src and xenocara release directories in order to fix the SHA256
shipped with installXX.iso. This also stops recomputing the SHA over
again, it's only performed for the files missing a SHA.

This is a breaking change as reldir is renamed to rel and the introduced
xenocara directory relx. The names are taken from
distrib/amd64/iso/Makefile.