rework regress configuration
Instead of accepting a list of test suites, make the regress keyword accept a
single test suite optionally followed by additional variables that influences
the execution. The previous design of adding flags after the path was a
limitation from back when the configuration was parsed using ksh(1).

While here, add support for adding additional environment variables per test
suite. This was the prime motivation for this breaking change.